+++ StoatApplication/Core/WebSocket/Models/ServerToClient/Pong.cs
@@ -3,7 +3,7 @@ using System.Text.Json.Serialization;
 namespace StoatApplication.Core.WebSocket.Models.ServerToClient;
 
 public sealed record Pong(
-    [property: JsonPropertyName("data")] int Data
+    [property: JsonPropertyName("data")] long Data
 ) : IEvent
 {
     public string Type => "Pong";
